Hide ending of all files once new file is added
On every new file addition that answers the name
pattern, a new tail command is issued. This causes the ending of all
files to be shown on screen. -n 0 will restart the tail without showing
last lines. We may miss a few lines but this will prevent junk on screen
when a lot of new files are created.